Fix bugs in centroid calculation of a cluster
1) Centroid should for all the terms in the current centroid
and in the incoming document, and not just the incoming document.

2) Use double instead of int in the DocumentVector since centroid
has floating-type term-counts.

+++ Indexer/src/cluster/DocumentVector.java
@@ -11,11 +11,11 @@ public class DocumentVector {
 
     private int docId;
 
-    private HashMap<String, Integer> docVector;
+    private HashMap<String, Double> docVector;
 
     public DocumentVector(int d) {
         docId = d;
-        docVector = new HashMap<String, Integer>();
+        docVector = new HashMap<String, Double>();
     }
 
     public int getDocId() {
@@ -24,18 +24,18 @@ public class DocumentVector {
 
     // add a count to an existing term in the vector
     public void addTerm(String term) {
-        docVector.put(term, docVector.getOrDefault(term, 0) + 1);
+        docVector.put(term, docVector.getOrDefault(term, 0.0) + 1.0);
     }
 
     // add a term with its count in the vector
-    public void addTerm(String term, int termFrequency) {
+    public void addTerm(String term, double termFrequency) {
         docVector.put(term, termFrequency);
     }
 
     public JSONObject JSONifySelf() {
         JSONObject jsonObject = new JSONObject();
-        for (Entry<String, Integer> entry : docVector.entrySet()) {
-            jsonObject.put(entry.getKey(), entry.getValue());
+        for (Entry<String, Double> entry : docVector.entrySet()) {
+            jsonObject.put(entry.getKey(), entry.getValue().intValue());
         }
         return jsonObject;
     }
@@ -49,8 +49,8 @@ public class DocumentVector {
         return docVector.keySet();
     }
 
-    public int getTermCount(String term) {
-        return docVector.getOrDefault(term, 0);
+    public Double getTermCount(String term) {
+        return docVector.getOrDefault(term, 0.0);
     }
 
 }
